leek: introduce support for AVX-512.
This new implementation brings support for AVX-512 capable CPUs.
This requires the BW extension which is necessary to handle byteswaps
in an efficient way. This extension is only supported by Xeon CPUs, as
opposed to KNL devices (could be worked around with reduced performances).

GCC-4 (even 4.9) will not be able to build AVX-512 implementation.
GCC-5 has some issues on CPUID detection for avx512bw (worked around).
Recommended compiler is thus GCC-6 or greater.